pitext_codegen/requirements.txt
fastapi>=0.110
uvicorn[standard]==0.24.0
openai>=1.14
python-dotenv>=1.0
